fix(preheat): add precheck when delete instance

+++ src/server/v2.0/handler/preheat.go
@@ -81,6 +81,22 @@ func (api *preheatAPI) DeleteInstance(ctx context.Context, params operation.Dele
 		return api.SendError(ctx, err)
 	}
 
+	// delete instance should check the instance whether be used by policies
+	policies, err := api.preheatCtl.ListPolicies(ctx, &q.Query{
+		Keywords: map[string]interface{}{
+			"provider_id": instance.ID,
+		},
+	})
+	if err != nil {
+		return api.SendError(ctx, err)
+	}
+
+	if len(policies) > 0 {
+		return api.SendError(ctx, liberrors.New(nil).
+			WithCode(liberrors.PreconditionCode).
+			WithMessage("Can't delete instance %s, %d preheat policies use it as provider", instance.Name, len(policies)))
+	}
+
 	err = api.preheatCtl.DeleteInstance(ctx, instance.ID)
 	if err != nil {
 		return api.SendError(ctx, err)
@@ -293,7 +309,7 @@ func (api *preheatAPI) DeletePolicy(ctx context.Context, params operation.Delete
 
 	// Detecting running tasks under the policy
 	if err = detectRunningExecutions(executions); err != nil {
-		return api.SendError(ctx, err)
+		return api.SendError(ctx, liberrors.New(err).WithCode(liberrors.PreconditionCode))
 	}
 
 	err = api.preheatCtl.DeletePolicy(ctx, policy.ID)
